Lab 4: Exercise 3: map per-CPU kernel stacks
FIXME:
In boot_map_region() we had to comment out 'v.a. is already mapped' assertion,
as it conflicts with CPU[0] kernel stack being first mapped in mem_init().
If we just remove the mapping in mem_init() - which currently looks as
a reasonable move, we hit a triple fault upon 1-st system call in the user-space
process.
Why ?
It is currently unclear.